services/bff/pyproject.toml
[tool.poetry]
name = "bff"
version = "0.1.0"
description = "Backend For Frontend - REST/WebSocket API gateway for Farmer Power Platform"
authors = ["Farmer Power Team"]
readme = "README.md"
packages = [{include = "bff", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
structlog = "^24.1.0"
tenacity = "^8.2.0"
httpx = "^0.26.0"
grpcio = "^1.60.0"
# OpenTelemetry - use compatible versions
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp = "^1.29.0"
opentelemetry-instrumentation-fastapi = ">=0.50b0"
# Shared libs
fp-common = { path = "../../libs/fp-common" }
fp-proto = { path = "../../libs/fp-proto" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = ">=0.26.0,<0.28.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
grpcio-testing = "^1.60.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

# Ruff config inherited from root pyproject.toml

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
